refactor(axelar-gateway)!: separate gateway messaging interface (#82)
