+++ contracts/axelar-gateway/src/executable.rs
@@ -1,7 +1,15 @@
 use axelar_soroban_std::ensure;
 use soroban_sdk::{contractclient, Address, Bytes, Env, String};
 
-use crate::{error::ContractError, AxelarGatewayClient};
+use crate::AxelarGatewayMessagingClient;
+use soroban_sdk::contracterror;
+
+#[contracterror]
+#[derive(Copy, Clone, Debug, Eq, PartialEq, PartialOrd, Ord)]
+#[repr(u32)]
+pub enum ExecutableError {
+    NotApproved = 1,
+}
 
 /// Interface for an Axelar Executable app.
 #[contractclient(name = "AxelarExecutableClient")]
@@ -9,7 +17,7 @@ pub trait AxelarExecutableInterface {
     /// Return the trusted gateway contract id.
     fn gateway(env: &Env) -> Address;
 
-    /// Execute a cross-chain contract call with the given payload. This function must validate that the contract call is received from the trusted gateway.
+    /// Execute a cross-chain message with the given payload. This function must validate that the message is received from the trusted gateway.
     fn execute(
         env: Env,
         source_chain: String,
@@ -18,19 +26,19 @@ pub trait AxelarExecutableInterface {
         payload: Bytes,
     );
 
-    /// Validate if a gateway has approved a contract call.
+    /// Validate if a gateway has approved a message.
     /// This should be called from an implementation of `execute` before executing custom app logic.
     /// This method doesn't get exposed from the contract, as Soroban SDK's contractimpl macro ignores default trait methods.
-    fn validate(
+    fn validate_message(
         env: &Env,
         source_chain: &String,
         message_id: &String,
         source_address: &String,
         payload: &Bytes,
-    ) -> Result<(), ContractError> {
-        let gateway = AxelarGatewayClient::new(env, &Self::gateway(env));
+    ) -> Result<(), ExecutableError> {
+        let gateway = AxelarGatewayMessagingClient::new(env, &Self::gateway(env));
 
-        // Validate the contract call was approved by the gateway
+        // Validate that the message was approved by the gateway
         ensure!(
             gateway.validate_message(
                 &env.current_contract_address(),
@@ -39,7 +47,7 @@ pub trait AxelarExecutableInterface {
                 source_address,
                 &env.crypto().keccak256(payload).into(),
             ),
-            ContractError::NotApproved
+            ExecutableError::NotApproved
         );
 
         Ok(())
