+++ contracts/axelar-gateway/src/lib.rs
@@ -6,9 +6,11 @@
 extern crate std;
 
 pub mod error;
-pub mod types;
-
 pub mod executable;
+mod messaging_interface;
+pub mod types;
+pub use messaging_interface::{AxelarGatewayMessagingClient, AxelarGatewayMessagingInterface};
+
 mod interface;
 
 #[cfg(all(target_family = "wasm", feature = "testutils"))]
