+++ integration-tests/tests/gmp.rs
@@ -23,7 +23,7 @@ impl AxelarExecutableInterface for AxelarApp {
         source_address: String,
         payload: Bytes,
     ) {
-        let _ = Self::validate(&env, &source_chain, &message_id, &source_address, &payload);
+        let _ = Self::validate_message(&env, &source_chain, &message_id, &source_address, &payload);
 
         env.events()
             .publish((Symbol::new(&env, "executed"),), (payload,));
